requirements.txt
boto3
certifi==2018.11.29
chardet==3.0.4
cognitive-face==1.4.2
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.9.0
drest==0.9.12
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
imutils==0.5.2
numpy==1.15.4
opencv-python==3.4.4.19
psycopg2==2.7.6.1
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
whitenoise==4.1.2
